+++ studio/stencil.config.ts
@@ -8,15 +8,13 @@ import replace from 'rollup-plugin-replace';
 
 // @ts-ignore
 const dev: boolean = process.argv && process.argv.indexOf('--dev') > -1;
+// @ts-ignore
+const staging: boolean = process.argv && process.argv.indexOf('--staging') > -1;
 
 const globalScript: string = dev ? 'src/global/app-dev.ts' : 'src/global/app.ts';
 
-// @ts-ignore
-import devConfig from './config.dev.json';
-// @ts-ignore
-import prodConfig from './config.prod.json';
-
-const configValues = dev ? devConfig : prodConfig;
+const configDataFile = dev ? (staging ? './config.staging.json' : './config.dev.json') : './config.prod.json';
+const configValues = require(configDataFile);
 
 export const config: Config = {
     outputTargets: [{