+++ webcomponents/inline-editor/src/components/inline-editor/deckdeckgo-inline-editor.tsx
@@ -315,12 +315,6 @@ export class DeckdeckgoInlineEditor {
         return;
       }
 
-      // Quirk when use click at the begin of the selection after having already selected something
-      if (this.selection && selection.toString() === this.selection.toString()) {
-        resolve();
-        return;
-      }
-
       const activated: boolean = await this.activateToolbar(selection);
       await this.setToolsActivated(activated);
 
@@ -785,14 +779,14 @@ export class DeckdeckgoInlineEditor {
         const textAfter: string = (index + this.anchorLink.text.length) > -1 ? target.textContent.substr((index + this.anchorLink.text.length)) : null;
 
         if (textBefore) {
-          target.parentElement.appendChild(document.createTextNode(textBefore));
+          target.parentElement.insertBefore(document.createTextNode(textBefore), target);
         }
 
         const a: HTMLAnchorElement = await this.createLinkElement();
-        target.parentElement.appendChild(a);
+        target.parentElement.insertBefore(a, target);
 
         if (textAfter) {
-          target.parentElement.appendChild(document.createTextNode(textAfter));
+          target.parentElement.insertBefore(document.createTextNode(textAfter), target);
         }
 
         target.parentElement.removeChild(target);