+++ webcomponents/inline-editor/CHANGELOG.md
@@ -1,3 +1,17 @@
+<a name="1.0.0-rc.3-2"></a>
+# 1.0.0-rc.3-2 (2019-10-14)
+
+### Fix
+
+* on mouse selection the link were not correctly created as it kept a selection reference to the first or last character ([ #410](https://github.com/deckgo/deckdeckgo/issues/410)
+
+<a name="1.0.0-rc.3-1"></a>
+# 1.0.0-rc.3-1 (2019-10-13)
+
+### Fix
+
+* create link don't always respect order ([#405](https://github.com/deckgo/deckdeckgo/issues/405))
+
 <a name="1.0.0-rc.3"></a>
 # 1.0.0-rc.3 (2019-10-01)